olf_2023/requirements.txt
jupyterlab
rise
jupysql
psycopg2
cogapp
records
pyperclip
pymongo
pandas
tabulate
jupysql
pugsql
agnostic[postgres]
agnostic[sqlite]
datasette
pypika
